feat: extra decompress into standalone bin
