+++ .gitignore
@@ -13,7 +13,10 @@ Cargo.lock
 # MSVC Windows builds of rustc generate these, which store debugging information
 *.pdb
 .idea/
-ryot_compass/assets/assets/
+assets/cip-catalog
+assets/sprite-sheets
+assets/catalog-content.json
+assets/appearances.dat
 
 .DS_Store
 .vscode/
