+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["ryot", "ryot_compass"]
+members = ["decompress", "ryot", "ryot_compass"]
 resolver = "2"
 
 # Enable a small amount of optimization in debug mode
