--- assets
@@ -1,1 +1,0 @@
-ryot_compass/assets
