+++ decompress/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "decompress"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+ryot = { version = "0.1.0", path = "../ryot", default-features = false, features = [
+    "compression",
+] }
+config = "0.13"
+log = "0.4"
+
+color-eyre.workspace = true
+glam.workspace = true
+image.workspace = true
+rayon.workspace = true
+serde.workspace = true
+
+[dev-dependencies]
+rstest.workspace = true
