+++ ryot/src/content.rs
@@ -1,14 +1,12 @@
 use std::sync::OnceLock;
 
 use crate::{error, SpriteSheetConfig};
-use config::Config;
 use glam::UVec2;
 use serde::Deserialize;
+use std::fs;
 use std::io::Read;
-use std::path::{Path, PathBuf};
-use std::{fs, result};
+use std::path::PathBuf;
 
-pub static CONTENT_CONFIG_PATH: &str = "config/.content.toml";
 pub static DYNAMIC_ASSETS_PATH: &str = "dyanmic.assets.ron";
 pub static SPRITE_SHEET_FOLDER: &str = "sprite-sheets";
 
@@ -46,48 +44,10 @@ pub fn assets_root_path() -> PathBuf {
     PathBuf::from("assets")
 }
 
-pub fn read_content_configs(config_path: PathBuf) -> ContentConfigs {
-    let settings = Config::builder()
-        .add_source(config::File::from(config_path))
-        .build()
-        .expect("Failed to build config")
-        .try_deserialize::<ContentConfigs>()
-        .expect("Failed to deserialize config");
-
-    let dir_settings = &settings.directories;
-
-    match is_path_within_root(&dir_settings.destination_path, Path::new("assets")) {
-        Ok(true) => settings,
-        Ok(false) | Err(_) => panic!(
-            "Target path {} is not within assets folder",
-            dir_settings
-                .destination_path
-                .to_str()
-                .expect("Failed to convert target path to str")
-        ),
-    }
-}
-
-pub fn is_path_within_root(
-    destination_path: &Path,
-    root_path: &Path,
-) -> result::Result<bool, std::io::Error> {
-    Ok(fs::canonicalize(destination_path)?.starts_with(fs::canonicalize(root_path)?))
-}
-
 pub fn get_full_file_buffer(path: &PathBuf) -> error::Result<Vec<u8>> {
     let mut file = fs::File::open(path)?;
     let mut buffer: Vec<u8> = Vec::new();
     file.read_to_end(&mut buffer)?;
 
     Ok(buffer)
-}
-
-type Result<T> = result::Result<T, config::ConfigError>;
-
-pub fn config_from<'de, T: Deserialize<'de>>(config_path: &str) -> Result<T> {
-    Config::builder()
-        .add_source(config::File::with_name(config_path))
-        .build()?
-        .try_deserialize::<T>()
 }
