+++ ryot/src/lib.rs
@@ -48,7 +48,6 @@ pub use content::*;
 #[cfg(feature = "lmdb")]
 pub mod lmdb;
 
-mod build;
 pub mod sprites;
 
 pub use sprites::*;
@@ -61,7 +60,6 @@ pub mod helpers;
 pub mod prelude {
     #[cfg(feature = "bevy")]
     pub use crate::bevy_ryot::*;
-    pub use crate::build::*;
     #[cfg(feature = "compression")]
     pub use crate::compression::{compress, decompress, Compression, Zstd};
     pub use crate::content::*;
