--- ryot_compass/assets/fixtures/.test.toml
@@ -1,1 +1,0 @@
-test = "i am a test"
