+++ ryot_compass/assets/sprite-sheets
@@ -1,0 +1,1 @@
+../../assets/sprite-sheets
